requirements_enhanced.txt
streamlit==1.55.0
pandas==2.3.3
PyPDF2==3.0.1
python-dotenv==1.2.2
openpyxl==3.1.5
langchain==1.2.12
langchain-openai==1.1.11
langchain-community==0.4.1
numpy==2.4.3
google-generativeai==0.8.6
